+++ packages/plugin-packages/mashroom-vhost-path-mapper/src/middleware/mashroom-bootstrap-middleware.ts
@@ -1,11 +1,14 @@
 
+import context from '../context';
 import MashroomVHostPathMapperMiddleware from './MashroomVHostPathMapperMiddleware';
 
 import type {MashroomMiddlewarePluginBootstrapFunction} from '@mashroom/mashroom/type-definitions';
 
 const bootstrap: MashroomMiddlewarePluginBootstrapFunction = async (pluginName, pluginConfig) => {
     const { considerHttpHeaders, hosts } = pluginConfig;
-    const middleware = new MashroomVHostPathMapperMiddleware(considerHttpHeaders, hosts);
+    context.considerHttpHeaders = considerHttpHeaders;
+    context.vhostDefinitions = hosts;
+    const middleware = new MashroomVHostPathMapperMiddleware();
     return middleware.middleware();
 };