+++ packages/plugin-packages/mashroom-vhost-path-mapper/src/utils/find_host_definition.ts
@@ -5,20 +5,20 @@ import type {
     VHostDefinitions,
 } from '../../type-definitions/internal';
 
-export default (determinedHost: DeterminedHost, hostDefinitions: VHostDefinitions): VHostDefinition | undefined => {
-    if (!hostDefinitions) {
+export default (determinedHost: DeterminedHost, vhostDefinitions: VHostDefinitions): VHostDefinition | undefined => {
+    if (Object.keys(vhostDefinitions).length === 0) {
         return undefined;
     }
 
-    const hostNames = Object.keys(hostDefinitions);
+    const hostNames = Object.keys(vhostDefinitions);
     for (let i = 0; i < hostNames.length; i ++) {
         const hostname = hostNames[i];
         if (hostname.indexOf(':') !== -1) {
            if (determinedHost.port && hostname === `${determinedHost.hostname}:${determinedHost.port}`) {
-                return hostDefinitions[hostname];
+                return vhostDefinitions[hostname];
             }
         } else if (hostname === determinedHost.hostname) {
-            return hostDefinitions[hostname];
+            return vhostDefinitions[hostname];
         }
     }