+++ packages/plugin-packages/mashroom-vhost-path-mapper/src/utils/map_path.ts
@@ -16,9 +16,11 @@ export default (originalUrl: string, vHostDefinition: VHostDefinition, reverse =
     if (mapping) {
         const determinedRules = reverse ? reverseRules(mapping) : mapping;
         const basePaths = Object.keys(determinedRules);
+        let matchingMapping = false;
         for (let i = 0; i < basePaths.length; i ++) {
             const basePath = basePaths[i];
             if (originalPath.startsWith(basePath)) {
+                matchingMapping = true;
                 const replacementPath = determinedRules[basePath];
                 const fixedReplacementPath = replacementPath === '/' ? '' : replacementPath;
                 const url = basePath === '/' && originalPath !== '/' ? replacementPath + originalUrl : originalUrl.replace(basePath, fixedReplacementPath);
@@ -50,6 +52,21 @@ export default (originalUrl: string, vHostDefinition: VHostDefinition, reverse =
                     };
                 }
             }
+        }
+        // If there is no matching mapping we still need to consider the frontendBasePath
+        if (!matchingMapping && reverse && vHostDefinition.frontendBasePath && vHostDefinition.frontendBasePath != '/') {
+            const frontendUrl = `${vHostDefinition.frontendBasePath}${originalUrl}`;
+            const frontendPath = frontendUrl.split(/[?#]/)[0];
+            return {
+                url: frontendUrl,
+                info: {
+                    mappingRuleBasePath: '',
+                    originalUrl,
+                    frontendBasePath,
+                    frontendPath,
+                    frontendUrl,
+                },
+            };
         }
     }