+++ packages/plugin-packages/mashroom-vhost-path-mapper/type-definitions/index.d.ts
@@ -19,17 +19,21 @@ export type RequestVHostMappingInfo = {
      */
     readonly frontendPath: string;
     /**
-     * The full frontend url seen by the user (browser url)
+     * The full frontend url seen by the user (including query and so on)
      */
     readonly frontendUrl: string;
 }
 
 export interface MashroomVHostPathMapperService {
+    /**
+     * Reverse map the given server url to the url as seen by the user (browser).
+     * The given URL must not contain host, only path with query params and so on.
+     */
+    getFrontendUrl(req: Request, url: string): string;
 
     /**
      * Get the details if the url of the current path has been rewritten
      */
-    getMappingInfo(request: Request): RequestVHostMappingInfo | undefined;
-
+    getMappingInfo(req: Request): RequestVHostMappingInfo | undefined;
 }